feat(hiresfix): HiResFix StaticPy 节点 + sd_pipeline_generate_hires C API + TYPE_MAP float 修复
- C API: sd_pipeline_load_ex (GGUF+LLM 支持), sd_compute_hires_resolution,
  sd_pipeline_generate_hires (base 计算 → 生成 → postproc → 保存)
- StaticPy 节点: DiffusionModelLoader (GGUF), HiResFix (目标分辨率,
  hires steps/strength, FreeU, clarity/sharpen postproc)
- postproc 库链接到 libsdcpp_adapter.so
- 修复 TYPE_MAP float→double 导致 float 参数传入 C 端被截断的 bug
- 新增 comfycli_remote.sh / xgc_ctl.py / remote_server.md 远程部署脚本

+++ staticpy/static_translate.py
@@ -44,7 +44,7 @@ def scheme_name(name):
 # ====== C 类型映射 ======
 TYPE_MAP = {
     "int": "int",
-    "float": "double",
+    "float": "float",
     "double": "double",
     "bool": "boolean",
     "str": "string",
